Rename api-keys.ts to api-key.ts
